+++ deltafi-core/src/test/java/org/deltafi/core/Util.java
@@ -136,6 +136,7 @@ public class Util {
         Assertions.assertThat(actual.getProtocolStack()).isEqualTo(expected.getProtocolStack());
         Assertions.assertThat(actual.getDomains()).isEqualTo(expected.getDomains());
         Assertions.assertThat(actual.getIndexedMetadata()).isEqualTo(expected.getIndexedMetadata());
+        Assertions.assertThat(actual.getIndexedMetadataKeys()).isEqualTo(expected.getIndexedMetadataKeys());
         Assertions.assertThat(actual.getEnrichment()).isEqualTo(expected.getEnrichment());
         Assertions.assertThat(actual.getFormattedData()).isEqualTo(expected.getFormattedData());
         Assertions.assertThat(actual.getEgressed()).isEqualTo(expected.getEgressed());